services/collector/requirements.txt
moomoo-api>=1.0.0
pandas>=2.0.0
redis>=5.0.0
pyyaml>=6.0.0
httpx>=0.27.0
